delete messageView in components and fix bug modal when closing
